requirements.txt
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
alignment==1.0.10
astor==0.8.1
async-timeout==3.0.1
attrs==21.2.0
autograd==1.3
autograd-gamma==0.5.0
blessings==1.7
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
colorful==0.5.4
cycler==0.10.0
editdistance==0.5.3
filelock==3.0.12
formulaic==0.2.4
future==0.18.2
google-api-core==1.31.3
google-auth==1.35.0
googleapis-common-protos==1.53.0
gower==0.0.5
gpustat==0.6.0
grpcio==1.40.0
hiredis==2.0.0
idna==3.2
interface-meta==1.2.4
joblib==1.0.1
jsonschema==3.2.0
kiwisolver==1.3.2
lifelines==0.26.3
llvmlite==0.34.0
matplotlib==3.4.3
msgpack==1.0.2
multidict==5.1.0
numba==0.51.2
numpy==1.20.3
nvidia-ml-py3==7.352.0
opencensus==0.7.13
opencensus-context==0.1.2
packaging==21.0
pandas==1.2.4
Pillow==8.3.2
prometheus-client==0.11.0
protobuf==3.17.3
psutil==5.8.0
py-spy==0.3.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynndescent==0.5.4
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
ray==1.6.0
redis==3.5.3
requests==2.26.0
rsa==4.7.2
sas7bdat==2.2.3
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
setproctitle==1.2.2
six==1.16.0
sklearn==0.0
threadpoolctl==2.2.0
typing-extensions==3.10.0.2
umap-learn==0.5.1
urllib3==1.26.7
wrapt==1.12.1
yarl==1.6.3
